% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{computeAPAIndex}
\alias{computeAPAIndex}
\title{Compute APA index}
\usage{
computeAPAIndex(APA, barcodes = NULL, method = "RUD")
}
\arguments{
\item{APA}{A PACdataset.}

\item{barcodes}{Selected cells or spots.}

\item{method}{Can be WUL, RUD, SLR. WUL: Weighted UTR length (Jia 2017; Ultisky) (3UTRAPA, choose2PA=NULL/PD/MOST) RUD: Relative Usage of Distal Poly(A) Sites (Ji 2009) (3UTRAPA with/without non-3UTR PACs, choose2PA=NULL/PD/MOST) only calc distal ratio, RUD=distal/gene SLR: short to long ratio (Begik 2017) (3UTRAPA, choose2PA=PD/most) #only for PD, SLR=short/long GPI: geometric proximal index (Shulman 2019) (3UTRAPA, choose2PA=NULL/PD/most) #GPI=proximal/geo_mean_PD}
}
\description{
Calculates APA index by different metrics for each condition.
}
