% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{computeIndex}
\alias{computeIndex}
\title{Cluster evaluation of different indexes}
\usage{
computeIndex(stAPAminerObj)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}
}
\description{
Cluster evaluation of different indexes,Cluster evaluation of different indexes,
including 4 internal indexes(DBI, CH, SC, Denn) and 4 external indexes(ARI, jaccard, purity, NMI).
}
