% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{computePearsonIndex}
\alias{computePearsonIndex}
\title{Compute Pearson correlation coefficient}
\usage{
computePearsonIndex(stAPAminerObj, source = "RUD")
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{source}{Source name for the object.}
}
\description{
Compute the Pearson correlation coefficient for each spot in the different layers.
}
