% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{convertToSCE}
\alias{convertToSCE}
\title{Convert Seurat object to SingleCellExperiment object}
\usage{
convertToSCE(stAPAminerObj)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}
}
\description{
Convert Seurat object to SingleCellExperiment object
}
