% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{createStAPAminerObject}
\alias{createStAPAminerObject}
\title{Create a stAPAminer object}
\usage{
createStAPAminerObject(count, metaData)
}
\arguments{
\item{count}{APA index}

\item{metaData}{Coordinate information for spatial transcriptome data}
}
\description{
Create a stAPAminer object from raw data.
}
