% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{drawClusterPlot}
\alias{drawClusterPlot}
\title{Draw the colored spatial scatter plot}
\usage{
drawClusterPlot(
  stAPAminerObj,
  title = NULL,
  labelOrder = levels(stAPAminerObj@seurat@active.ident),
  color = rainbow(cluster_nums),
  size = 3,
  islegend = TRUE
)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{title}{the title of the plot.}

\item{labelOrder}{the order of legend.}

\item{color}{The color level.}

\item{size}{The size of spot.}

\item{islegend}{whether to display legend.}
}
\description{
Graphs the output of a spatial scatter plot where each point is a spot and it's positioned based on the additional coordinate information.
By default, spots are colored by the active.ident.
}
