% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{drawSpatialExpress}
\alias{drawSpatialExpress}
\title{Visualize individual gene expression levels on a spatial scatter plot}
\usage{
drawSpatialExpress(stAPAminerObj, geneName, title = geneName, size = 2)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{geneName}{The gene name}

\item{title}{the title of the plot.}

\item{size}{the size of the plot.}
}
\description{
Colors single spots on a spatial scatter plot according to the gene expression.
}
