% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findAllMarkers}
\alias{findAllMarkers}
\title{APA index markers for all identity classes}
\usage{
findAllMarkers(stAPAminerObj, padj = 0.05, logFC = 0.5)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{padj}{Maximum adjusted p-value.}

\item{logFC}{Minimum log fold-change of the average expression between the two groups.}
}
\description{
Finds markers (differentially expressed APA index) for identity classes
}
