% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findContrastAPA}
\alias{findContrastAPA}
\title{APA index markers for pairs of identity classes}
\usage{
findContrastAPA(
  stAPAminerObj,
  padj = 0.05,
  logFC = 0.5,
  levels = c("GCL", "MCL", "OPL", "GL", "ONL")
)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{padj}{Maximum adjusted p-value.}

\item{logFC}{Minimum log fold-change of the average expression between the two groups.}

\item{levels}{Identity classes to be contrasted}
}
\description{
Finds markers (differentially expressed APA index) for identity classes
}
