% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findLabels}
\alias{findLabels}
\title{Spatial Scatter Plot}
\usage{
findLabels(stAPAminerObj, size = 4)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{size}{The size of spots.}
}
\description{
Graphs the output of a spatial scatter plot where each point is a spot and it's positioned based on the additional coordinate information.
By default, spots are colored by the active.ident.
}
