% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findstAPASet}
\alias{findstAPASet}
\title{Merging multiple types of differential APA genes}
\usage{
findstAPASet(stAPAminerObj)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}
}
\description{
Merging multiple types of differential APA genes
}
