% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findSVAPA}
\alias{findSVAPA}
\title{Find genes with spatial patterns of APA usage variation}
\usage{
findSVAPA(stAPAminerObj, padj = 0.05)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{padj}{Maximum adjusted p-value.}
}
\description{
we also adopted SPARK to detect genes with spatial patterns of APA usage variation. SVAPA genes mark distinct spatial APA usage patterns in the global spatial context.
}
