% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{imputeAPAIndex}
\alias{imputeAPAIndex}
\title{Imputation APA index}
\usage{
imputeAPAIndex(index, gene, k = 10, init = TRUE)
}
\arguments{
\item{index}{APA index to be recovered}

\item{gene}{reference gene expression matrix}

\item{k}{Defines k for the k-nearest neighbor algorithm}

\item{init}{Whether to initialize the APA index}
}
\description{
A KNN-based imputation model for recovering APA signals in the APA index matrix.
}
