% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{makeStCluster}
\alias{makeStCluster}
\title{Preprocessing and Clustering}
\usage{
makeStCluster(
  stAPAminerObj,
  nfeatures = 2000,
  dims = 1:5,
  resolution = 0.5,
  k = 20
)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{nfeatures}{Number of features to select as top variable features.}

\item{dims}{Dimensions of reduction to use as input.}

\item{resolution}{Value of the resolution parameter.}

\item{k}{Defines k for the k-nearest neighbor algorithm}
}
\description{
Create a stAPAminer object from raw data.
}
