% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R
\name{optimalKvalue}
\alias{optimalKvalue}
\title{Calculate the optimal k value}
\usage{
optimalKvalue(data, count, position, ncluster, orders = c(seq(4, 20, 2)))
}
\arguments{
\item{data}{APA index to be recovered}

\item{count}{reference gene expression matrix}

\item{position}{Location information of spots}

\item{ncluster}{The number of clusters in the target cluster}

\item{orders}{Which k values to test, we recommend from 4 to 20}
}
\description{
Calculate the clustering index after imputed under different k values,
and obtain the optimal k value through the comprehensive index
}
