% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{readSierra}
\alias{readSierra}
\title{Convert Sierra result files to movAPA object}
\usage{
readSierra(path, annodb)
}
\arguments{
\item{path}{path to Sierra result file. The path needs to contain peak_annotations.txt, counts files}

\item{annodb}{GTF file annotating poly(A) sites information}
}
\description{
Convert Sierra result files to movAPA object
}
