% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{RenameCluster}
\alias{RenameCluster}
\title{Rename the object's identity classes}
\usage{
RenameCluster(stAPAminerObj, layer_names = c("c1", "c2", "c3", "c4", "c5"))
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{layer_names}{The new object's identity classes}
}
\description{
Rename the object's identity classes.
}
