% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAPA.R
\name{findSpatialPattern}
\alias{findSpatialPattern}
\title{Identification of APA genes with spatial patterns}
\usage{
findSpatialPattern(stAPAminerObj, features, k = 10)
}
\arguments{
\item{stAPAminerObj}{A stAPAminer object.}

\item{features}{Genes to be tested with spatial patterns of variation in APA usage}

\item{k}{Number of spatial patterns}
}
\description{
Identification of APA genes with spatial patterns of APA usage variation
}
